Return a result when closing the channel, fixed naming issues.
As closing/aborting a channel might be a no-op if the channel is not open,
a corresponding result will be returned by the respective channel methods:

A return value of 'true' indicates that the channel is open and a request
to close/abort the channel will be sent.

A return value of 'false' (= no-op) indicates that no request will be sent
and, consequently, the specified response handler will NOT be invoked.